Help
Anyone who can help with this, you are more than welcome to do so! For now I'm just trying to get a qr code image to generate in Cate Test3.html, from code in printYoutubeTest.js. You may notice thre is a box on Cate Test3.html for the code, which is informed from printYoutube.css I believe. I'll have slack open and be checking it fairly frequently, and I'm down to make time to talk to one of you about it. I want to be clear I want help figuring out how to make this work but I don't want the issue to be solved for me, if that makes sense; I'll never learn if I'm magically presented with a solution. Thanks, and I look forward to talking soon.