Added alternate approach for debug override
Swapped to global variables for class declaration and instance declaration so that debug overrides production. LibreTexts.active check ensures code is only executed once.

This approach assumes that the load event listener function is identical in both the production and debug code. All modified functionality must instead be in the called class or function.